+++ packages/storage_controller/lib/storage_controller.dart
@@ -3,6 +3,7 @@ library storage_controller;
 import 'dart:typed_data';
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:hive/hive.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -13,7 +14,7 @@ import 'package:storage_controller/src/storage.dart';
 
 enum StorageMethod { HIVE, FLUTTER_SECURE_STORAGE, SHARED_PREFERENCES }
 
-class StorageController extends Storage {
+abstract class StorageController extends Storage {
   final _ensureInitializedMessage = '''
     [StorageController] must be initialized inside `runApp()`.
     
@@ -21,70 +22,66 @@ class StorageController extends Storage {
     
     If you think this is an error, please create an issue at: https://https://github.com/4itworks/opensource_qwkin_dart
   ''';
-  final StorageMethod method;
-
+  StorageMethod _method;
   bool _ensureInitialized = false;
   Storage _storage;
+
+  StorageMethod get method => _method;
+  bool get isInitialized => _ensureInitialized;
 
   @visibleForTesting
   Type get storageRuntimeType => _storage.runtimeType;
 
-  StorageController._(
-    this.method, {
-
-    /// flutter_secure_storage optional parameters
-    IOSOptions iosOptions,
-    AndroidOptions androidOptions,
-
-    /// hive optional parameters
+  StorageController.hive(
+    String storageName, {
     HiveCipher encryptionCipher,
     bool crashRecovery = true,
     String path,
     Uint8List bytes,
   }) {
-    switch (method) {
-      case StorageMethod.HIVE:
-        _initializeHive();
-        break;
-      case StorageMethod.FLUTTER_SECURE_STORAGE:
-        _initializeFlutterSecureStorage();
-        break;
-      case StorageMethod.SHARED_PREFERENCES:
-        _initializeSharedPreferencesStorage();
-        break;
-    }
+    _method = StorageMethod.HIVE;
+    _initializeHive(storageName,
+        encryptionCipher: encryptionCipher,
+        crashRecovery: crashRecovery,
+        path: path,
+        bytes: bytes);
   }
 
-  factory StorageController.hive({
+  StorageController.flutterSecureStorage(
+      {IOSOptions iosOptions, AndroidOptions androidOptions}) {
+    _method = StorageMethod.FLUTTER_SECURE_STORAGE;
+    _initializeFlutterSecureStorage(
+        iosOptions: iosOptions, androidOptions: androidOptions);
+  }
+  StorageController.sharedPreferences() {
+    _method = StorageMethod.SHARED_PREFERENCES;
+    _initializeSharedPreferencesStorage();
+  }
+
+  void _initializeHive(
+    String storageName, {
     HiveCipher encryptionCipher,
     bool crashRecovery = true,
     String path,
     Uint8List bytes,
-  }) =>
-      StorageController._(StorageMethod.HIVE,
-          encryptionCipher: encryptionCipher,
-          crashRecovery: crashRecovery,
-          path: path,
-          bytes: bytes);
-  factory StorageController.flutterSecureStorage(
-          {IOSOptions iosOptions, AndroidOptions androidOptions}) =>
-      StorageController._(StorageMethod.FLUTTER_SECURE_STORAGE,
-          iosOptions: iosOptions, androidOptions: androidOptions);
-  factory StorageController.sharedPreferences() =>
-      StorageController._(StorageMethod.SHARED_PREFERENCES);
-
-  void _initializeHive() async {
+  }) async {
     _storage = HiveBasedStorage();
 
-    await (_storage as HiveBasedStorage).initialize();
+    await (_storage as HiveBasedStorage).initialize(storageName,
+        encryptionCipher: encryptionCipher,
+        crashRecovery: crashRecovery,
+        path: path,
+        bytes: bytes);
 
     _ensureInitialized = true;
   }
 
-  void _initializeFlutterSecureStorage() {
+  void _initializeFlutterSecureStorage(
+      {IOSOptions iosOptions, AndroidOptions androidOptions}) {
     _storage = FlutterSecureStorageBasedStorage();
 
-    (_storage as FlutterSecureStorageBasedStorage).initialize();
+    (_storage as FlutterSecureStorageBasedStorage)
+        .initialize(iosOptions: iosOptions, androidOptions: androidOptions);
 
     _ensureInitialized = true;
   }
@@ -100,8 +97,8 @@ class StorageController extends Storage {
     WidgetsFlutterBinding.ensureInitialized();
 
     if (_storage is FlutterSecureStorageBasedStorage) {
-      // ignore: invalid_use_of_visible_for_testing_member
       (_storage as FlutterSecureStorageBasedStorage)
+          // ignore: invalid_use_of_visible_for_testing_member
           .setUpMockAndInitialize(flutterSecureStorage);
     }
     // ignore: invalid_use_of_visible_for_testing_member
@@ -109,6 +106,7 @@ class StorageController extends Storage {
   }
 
   @override
+  @mustCallSuper
   Future<void> delete(String key) async {
     assert(_ensureInitialized, _ensureInitializedMessage);
 
@@ -116,6 +114,7 @@ class StorageController extends Storage {
   }
 
   @override
+  @mustCallSuper
   Future<T> read<T>({String key}) async {
     assert(_ensureInitialized, _ensureInitializedMessage);
 
@@ -123,7 +122,7 @@ class StorageController extends Storage {
   }
 
   @override
-  // TODO: implement values
+  @mustCallSuper
   Future<Map<String, dynamic>> get values async {
     assert(_ensureInitialized, _ensureInitializedMessage);
 
@@ -131,6 +130,7 @@ class StorageController extends Storage {
   }
 
   @override
+  @mustCallSuper
   Future<void> wipe() async {
     assert(_ensureInitialized, _ensureInitializedMessage);
 
@@ -138,6 +138,7 @@ class StorageController extends Storage {
   }
 
   @override
+  @mustCallSuper
   Future<void> write<T>({String key, T value}) async {
     assert(_ensureInitialized, _ensureInitializedMessage);