+++ packages/storage_controller/test/hive_based_storage_test.dart
@@ -6,7 +6,7 @@ void main() {
 
   group('Hive storage tests', () {
     setUp(() async {
-      await storage.initialize();
+      await storage.initialize('myAwesomeStorage');
     });
 
     test('Should write and read with success', () async {